\name{metacoop-package}
\alias{metacoop-package}
\alias{metacoop}
\docType{package}
\title{
\packageTitle{metacoop}
}
\description{
\packageDescription{metacoop}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{metacoop}
\packageIndices{metacoop}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{metacoop}

Maintainer: \packageMaintainer{metacoop}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
